+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 02.01.2025
+- Input in user component (side branch 02_input)
 - Signals in user component (side branch 01_signals)
 - Improvement of user component
 - New user component